Create Graph Traversal Sub-Tutorial.md
This is a sub-tutorial of the main BMEG Beginner Tutorial. It was designed to fill-in more technical information that the main tutorial couldn't go into. In a way, it is an extension of the main tutorial. A link to this sub-tutorial's web page needs to be added to the main tutorial web page. 
It would also be useful to put a link of the main tutorial on the page of this tutorial so when people are done reading through the sub-tutorial, they can go back to the main tutorial. 
This document is a first draft. It needs more revision and even reformatting before being published. 

Developed by Mekdes Hilete